Transform graphics and input systems to singletons and add SpeechSystem
ConsoleGraphicsSystem and ConsoleInputSystem are now instantiated through GameGraphicsMode.
GameGraphicsMode define graphics for the game (for now, only console) and create corresponding
systems (can't have a ConsoleGraphicsSystem with a 3DInputSystem).

Add also a SpeechSystem that let messages be printed letter by letter in the console, with a
delay between them. The pause between each letter is made with the Sleeper helper.