Change system management in GameGraphicsMode to avoid blocking tests waiting for input
ConsoleInputSystem was instanciated and stored as an attribute in GameGraphicsMode for
console mode, but the same ConsoleInputSystem was returned no matter how many times
the Game was instantiated. This resulted in blocking tests waiting for input,
even when a simulated input was provided to the System.in

GameGraphicsMode now stores a reference to the ConsoleInputSystem class when in
console mode, and the getter for it is in charge of creating a new instance of
the class. As a result, when changing the System.in, we only need to get the instance
of the InputSystem from the GameGraphicsMode to update the input source and set it
to the Game.inputSystem